Cadastro Usuario funcional (cadastro)

+++ src/model/Cliente.java
@@ -3,7 +3,7 @@ package model;
 import java.util.Date;
 
 public class Cliente extends Usuario{
-    public Cliente(String nome, boolean sexo, Date nascimento, int cpf, int identidade, Endereco endereco) {
+    public Cliente(String nome, boolean sexo, Date nascimento, long cpf, long identidade, Endereco endereco) {
         super(nome, sexo, nascimento, cpf, identidade, endereco);
     }
 }
